server/requirements.txt
Flask
google-generativeai
python-dotenv
Flask-Cors
pymongo[srv]
gunicorn